# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: helmstudio platform API
  version: "1.0.0"
  description: >
    The local, single-machine API a studio's runtime SDK speaks, and the one
    the launcher speaks to the daemon. Served by the daemon at
    127.0.0.1:8700 and, for one studio with no daemon, by `helm dev` and the
    embedded provider (05 §2, §5a). The two are substitutable: one
    conformance suite runs against both (Q25). The in-process embedded
    provider has no token, and still enforces the studio's capabilities,
    read from its manifest, with the same 403 `capability_required`; the
    suite's capability cases run against both providers (M4 first review,
    #7).
servers:
  - url: http://127.0.0.1:8700/api/v1
...
security: []
paths:
...
  /launcher/jobs:
    get:
      tags: [launcher]
      operationId: launcherJobsList
      x-helm-group: jobs
      x-helm-method: list
      summary: Every studio's install, build, download, update, export and uninstall jobs, newest first.
      description: >
        No token. Never `task` jobs: those are a studio's own data and reach
        only that studio's token (Q7).
      parameters:
        - name: studio
          in: query
          schema: { $ref: "#/components/schemas/StudioIdValue" }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: A page of jobs.
          content: { application/json: { schema: { $ref: "#/components/schemas/JobPage" } } }
        "400": { $ref: "#/components/responses/Invalid" }
  /launcher/jobs/{id}:
    get:
      tags: [launcher]
      operationId: launcherJobsGet
      x-helm-group: jobs
      x-helm-method: get
      summary: One lifecycle job, with its steps. A task job is 404.
      parameters: [{ $ref: "#/components/parameters/JobId" }]
      responses:
        "200":
          description: The job.
          content: { application/json: { schema: { $ref: "#/components/schemas/Job" } } }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /launcher/jobs/{id}/logs:
    get:
      tags: [launcher]
      operationId: launcherJobsLogs
      x-helm-group: jobs
      x-helm-method: logs
      summary: A lifecycle job's build log as server-sent events. A task job is 404.
      description: Events as for GET /jobs/{id}/logs.
      parameters: [{ $ref: "#/components/parameters/JobId" }]
      x-helm-events:
        step: { $ref: "#/components/schemas/JobLogStep" }
        line: { $ref: "#/components/schemas/JobLogLine" }
        end: { $ref: "#/components/schemas/JobLogEnd" }
      responses:
        "200":
          description: An event stream.
          content: { text/event-stream: { schema: { type: string } } }
        "404": { $ref: "#/components/responses/NotFound" }

  # ======================================================================
  # Launcher: studios, install, supervision (01 §6-7; M2 and M3 as served)
  # ======================================================================
...
  /studios/{id}/selection:
    put:
      tags: [launcher]
      operationId: studiosSelect
      x-helm-group: studios
      x-helm-method: select
      summary: Choose which of a studio's selectable weights it launches with.
      description: >
        Q20, Q21. Exactly one selectable binding per studio is selected, which a
        partial unique index enforces rather than a check in Go. Changing it
        while the studio runs is 409 `conflict`; choosing a weight that is
        neither downloaded nor linked is 409 `not_fetched`, naming `:fetch` and
        `:link`. Changing it while stopped needs **no new approval** — every
        selectable weight was approved with the manifest.
      parameters: [{ $ref: "#/components/parameters/StudioId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/SelectionRequest" } } }
      responses:
        "200":
          description: The studio, with its new selection.
          content: { application/json: { schema: { $ref: "#/components/schemas/Studio" } } }
        "400": { $ref: "#/components/responses/Invalid" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }

  # ======================================================================
  # Launcher: writing manifests (R3a-R3d, 05 §5a; docs/decisions.md M7 Q17-Q19)
  #
  # The page never parses YAML. It sends text and gets back errors with lines,
  # the criteria and the document as JSON; a form edit goes as a JSON pointer
  # and a value, and the daemon applies it to the node tree. A second parser in
  # JavaScript would disagree with this one about anchors, merge keys, `on` and
  # duplicate keys — and would show one manifest while validating another.
  # ======================================================================
...
  /models:
    get:
      tags: [launcher]
      operationId: modelsList
      x-helm-group: models
      x-helm-method: list
      summary: Every model artifact, newest first, with the studios bound to it.
      parameters:
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: A page of artifacts.
          content: { application/json: { schema: { $ref: "#/components/schemas/ModelArtifactPage" } } }
        "400": { $ref: "#/components/responses/Invalid" }
  /models/{id}:
    get:
      tags: [launcher]
      operationId: modelsGet
      x-helm-group: models
      x-helm-method: get
      summary: One artifact, with the one-item reclaim preview a DELETE must confirm when it is an unused download.
      parameters: [{ $ref: "#/components/parameters/ModelId" }]
      responses:
        "200":
          description: The artifact.
          content: { application/json: { schema: { $ref: "#/components/schemas/ModelArtifactView" } } }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [launcher]
      operationId: modelsDelete
      x-helm-group: models
      x-helm-method: delete
      summary: Unlink a linked artifact, or reclaim an unused download with the confirm from its preview.
      description: >
        docs/decisions.md M3 review round 1 #3. A linked artifact: the symlink
        is removed, never the user's directory. A download: only with
        `confirm`; without it 409 `confirm_required`, with a stale one 409
        `preview_changed`, both carrying `details.preview`. 409 `in_use` names
        the studios bound to it; 409 `not_reclaimable` while it downloads.
      parameters:
        - { $ref: "#/components/parameters/ModelId" }
        - name: confirm
          in: query
          schema: { type: string }
      responses:
        "204": { description: "Unlinked or deleted." }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /models:reclaim:
    get:
      tags: [launcher]
      operationId: modelsReclaimPreview
      x-helm-group: models
      x-helm-method: reclaimPreview
      summary: Exactly which downloads reclaim would delete — managed artifacts no studio is bound to.
      responses:
        "200":
          description: The preview.
          content: { application/json: { schema: { $ref: "#/components/schemas/ModelReclaimPreview" } } }
    post:
      tags: [launcher]
      operationId: modelsReclaim
      x-helm-group: models
      x-helm-method: reclaim
      summary: Delete exactly the previewed set, or nothing (409 preview_changed with the current preview).
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ReclaimConfirm" } } }
      responses:
        "200":
          description: What was deleted.
          content: { application/json: { schema: { $ref: "#/components/schemas/ModelReclaimPreview" } } }
        "400": { $ref: "#/components/responses/Invalid" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    studioToken:
      type: http
      scheme: bearer
      description: >
        HELM_TOKEN, minted per launch for one studio and its declared
        capabilities, revoked when its group stops (07 §3, Q6). Format
        `hs_live_` plus 52 base32 characters.
...
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      description: The `next_cursor` of the previous page, unchanged.
      schema: { type: string }
...
    StudioId:
      name: id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/StudioIdValue" }
...
    ModelId:      { name: id, in: path, required: true, schema: { $ref: "#/components/schemas/ULID" } }
...
    JobId:        { name: id, in: path, required: true, schema: { $ref: "#/components/schemas/ULID" } }
...
  responses:
    Invalid:
      description: >
        The request is malformed. Codes: `bad_request`, `bad_filter`,
        `bad_cursor`, `invalid_document`.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    NotFound:
      description: Absent, deleted, or not the caller's to see. Code `not_found`.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Conflict:
      description: >
        Codes: `etag_mismatch`, `name_taken`, `job_finished`,
        `already_running`, `not_running`, `port_conflict`, `heavy_conflict`
        (with `details.heavy`), `not_installed`, `conflict`, `in_use`,
        `missing`, `confirm_required` and `preview_changed` (with
        `details.preview`), `not_reclaimable`, and from M7:
        `approval_required` (with `details.approval`), `id_taken`,
        `not_fetched`, `already_exists`.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Unprocessable:
      description: >
        Well-formed but refused. Codes: `outside_roots`, `cross_device`,
        `not_a_file`, `unknown_session`, `self_handoff`, `not_launchable`,
        `blocked`, `not_linkable`.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    ULID:
      type: string
      pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
    StudioIdValue:
      type: string
      pattern: "^[a-z][a-z0-9-]{1,38}[a-z0-9]$"
      description: A manifest id (schema/manifest.json `id`).
    JSONObject:
      type: object
      additionalProperties: true
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string, description: "A stable snake_case code a program branches on." }
        message: { type: string, description: "A sentence for a person." }
        details: { type: object, additionalProperties: true }
...
    Page:
      type: object
      required: [items, next_cursor]
      properties:
        next_cursor:
          type: [string, "null"]
          description: Null on the last page.

    # ---------------------------------------------------------------- identity
...
    ReclaimConfirm:
      type: object
      required: [confirm]
      additionalProperties: false
      properties:
        confirm: { type: string }

    # ---------------------------------------------------------------- gallery
...
    JobKind:
      type: string
      enum: [install, build, download, update, export, uninstall, task]
    JobState:
      type: string
      enum: [queued, running, succeeded, failed, cancelled, interrupted]
    JobError:
      type: object
      required: [code, message]
      description: >
        installations.last_failure and jobs.last_error (02 §4). Install codes:
        clone_failed, tool_missing, env_failed, step_failed, step_timeout,
        cancelled, interrupted, auth_required, disk_space, weights_failed,
        linked_missing, uninstall_failed. env_failed (creating a Python
        environment, docs/decisions.md M5 Q7) has no step_index. A task job's
        code is the studio's own.
      properties:
        code: { type: string }
        message: { type: string }
        phase: { type: string }
        step_index: { type: integer }
        exit_code: { type: integer }
        log_file_id: { $ref: "#/components/schemas/ULID" }
    StepRun:
      type: object
      required: [step_index, step_name, command, state]
      properties:
        step_index: { type: integer }
        step_name: { type: string }
        command: { type: string }
        state: { type: string, enum: [pending, running, succeeded, skipped, failed, interrupted, cancelled] }
        exit_code: { type: integer }
        log_file_id: { $ref: "#/components/schemas/ULID" }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
    Job:
      type: object
      required: [id, kind, state, progress_num, progress_den, created_at]
      properties:
        id: { $ref: "#/components/schemas/ULID" }
        kind: { $ref: "#/components/schemas/JobKind" }
        studio_id: { $ref: "#/components/schemas/StudioIdValue" }
        state: { $ref: "#/components/schemas/JobState" }
        subject_kind: { type: string }
        subject_id: { type: string }
        progress_num: { type: integer, minimum: 0 }
        progress_den: { type: integer, minimum: 0 }
        last_error: { $ref: "#/components/schemas/JobError" }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        cancel_requested_at: { type: string, format: date-time }
        steps: { type: array, items: { $ref: "#/components/schemas/StepRun" } }
        warnings: { type: array, items: { type: string } }
    JobPage:
      allOf:
        - { $ref: "#/components/schemas/Page" }
        - properties:
            items: { type: array, items: { $ref: "#/components/schemas/Job" } }
...
    JobLogStep:
      type: object
      required: [step_index, step_name, command]
      properties:
        step_index: { type: integer }
        step_name: { type: string }
        command: { type: string }
    JobLogLine:
      type: object
      required: [text]
      properties:
        step_index: { type: integer, description: "Absent for a task job." }
        text: { type: string }
    JobLogEnd:
      type: object
      required: [state]
      properties:
        state: { $ref: "#/components/schemas/JobState" }
        last_error: { $ref: "#/components/schemas/JobError" }

    # ---------------------------------------------------------------- studios

    # ---------------------------------------------------------------- timeline
...
    InstallState:
      type: string
      enum: [listed, cloning, cloned, building, built, fetching_weights, auth_required, ready, update_available, failed_clone, failed_build, failed_weights, removing]
    Studio:
      type: object
      required: [manifest_loaded, id, name, kinds, heavy, root, root_present, group]
      properties:
        manifest_loaded: { type: boolean, description: "False for a studio still running from a previous daemon whose manifest is gone or invalid. It can be stopped, not launched." }
        id: { $ref: "#/components/schemas/StudioIdValue" }
        name: { type: string }
        description: { type: string }
        kinds: { type: array, items: { type: string, enum: [video, image, audio, text] } }
        heavy: { type: boolean }
        peak_ram_gb: { type: integer }
        root: { type: string, description: "The recorded checkout, or where install will put it." }
        root_present: { type: boolean }
        group: { $ref: "#/components/schemas/ProcessGroup" }
        install_state: { $ref: "#/components/schemas/InstallState" }
        root_path: { type: string }
        commit_sha: { type: string }
        rebuild_needed: { type: boolean, description: "The manifest changed since this checkout was built from it." }
        runtime_env: { $ref: "#/components/schemas/JSONObject" }
        last_failure: { $ref: "#/components/schemas/JobError" }
        size_bytes: { type: integer }
        job_id: { $ref: "#/components/schemas/ULID" }

        # The library (M7 Q6, Q8, Q15, Q18, Q21). An entry exists for every id
        # the machine knows about, installed or not, valid or not.
        source: { $ref: "#/components/schemas/LibrarySource" }
        overrides:
          $ref: "#/components/schemas/LibrarySource"
          description: "The source this entry shadows, when it shadows one. What the card means by \"overrides the registry entry\", and what Revert would go back to."
        level: { $ref: "#/components/schemas/CertificationLevel" }
        manifest_state: { $ref: "#/components/schemas/ManifestState" }
        manifest_valid: { type: boolean }
        errors:
          type: array
          items: { $ref: "#/components/schemas/ManifestError" }
          description: "Why an invalid entry is invalid. It is listed rather than hidden (R2, amended M7 Q6), so a typo in an override does not make a studio vanish."
        manifest_file: { type: string, description: "Absolute path, when the manifest is a file on this machine." }
        provenance: { $ref: "#/components/schemas/Provenance" }
        repo: { type: string }
        ref: { type: string }
        local_path: { type: string }
        approval_required: { type: boolean, description: "True when the recorded approval does not match what would run now, so the next install, retry or launch will ask (Q10)." }
        selection: { type: string, description: "The chosen `selectable` weight, when the studio declares any. A launch without one is refused (Q21)." }
        selectable:
          type: array
          items: { $ref: "#/components/schemas/ApprovalWeight" }
          description: "The checkpoints this studio can be launched with, and which are already downloaded or linked."
        rebuild_needed_reason: { type: string, description: "Set when `rebuild_needed`: the manifest changed since this checkout was built from it, which an Override causes." }
        hue:
          $ref: "#/components/schemas/StudioHue"
          description: "The identity hue a launcher draws the studio's stripe and dot in. Always present, including for an entry whose manifest did not load (2026-09-17, the launcher redesign)."
...
    LibrarySource:
      type: string
      enum: [local, repo, registry, running]
      description: >
        Where this entry's manifest came from, highest first (Q6). `local` is
        `<data>/studios/<id>.yaml`; `repo` is `helmstudio.yaml` in the studio's
        own repository at the pinned ref, or in its installed checkout;
        `registry` is the inline manifest a pointer carries. `running` is a
        studio still running from a previous daemon whose manifest is gone.
...
    SelectionRequest:
      type: object
      additionalProperties: false
      required: [weight]
      properties:
        weight: { type: string, description: "The manifest name of a `selectable` weight." }
...
    ProcessGroup:
      type: object
      required: [studio_id, processes]
      properties:
        studio_id: { $ref: "#/components/schemas/StudioIdValue" }
        group_run_id: { $ref: "#/components/schemas/ULID" }
        state: { type: string, enum: [starting, running, stopping, failed, exited] }
        processes: { type: array, items: { $ref: "#/components/schemas/Process" } }
        failure: { $ref: "#/components/schemas/GroupFailure" }
...
    ModelArtifactView:
      allOf:
        - { $ref: "#/components/schemas/ModelArtifact" }
        - properties:
            reclaim: { $ref: "#/components/schemas/ModelReclaimPreview" }
    ModelArtifactPage:
      allOf:
        - { $ref: "#/components/schemas/Page" }
        - properties:
            items: { type: array, items: { $ref: "#/components/schemas/ModelArtifact" } }
...
    ModelReclaimPreview:
      type: object
      required: [items, total_bytes, confirm]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/ModelReclaimItem" } }
        total_bytes: { type: integer }
        confirm: { type: string }
...